% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_tables.R
\name{add_labels_as_columns}
\alias{add_labels_as_columns}
\title{Add labels with values as columns}
\usage{
add_labels_as_columns(data, labels)
}
\arguments{
\item{data}{a list of dataframes}

\item{labels}{strings that identify the datasets}
}
\value{
the data with the labels and their values appended as
  columns
}
\description{
Add Labels as Columns
}
\details{
This function adds labels as columns to each dataframe in a list
of dataframes. This may be useful when datasets can be combined
into one dataframe.
}
\author{
Jones
}
