% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_tables.R
\name{annotate_and_combine}
\alias{annotate_and_combine}
\title{Annotate and combine dataframes}
\usage{
annotate_and_combine(data, labels = NULL, combos = NULL)
}
\arguments{
\item{data}{a list of dataframes}

\item{labels}{a vector of strings that identify the datasets}

\item{combos}{a list of vectors of named logical values that
identify the sets of labels that should ultimately be included
or excluded}
}
\value{
a collection of tables to be analyzed
}
\description{
Annotate and Combine
}
\details{
Annotate and combine all of the dataframes in your list according
your rules. If a list of combinations is provided, you need not
provide the labels field. If no list of combintions is provided
then each label is assumed to become a group.
}
\author{
Jones
}
