% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_tables.R
\name{fun_plot}
\alias{fun_plot}
\title{A function to plot}
\usage{
fun_plot(
  data,
  x = "reporter",
  y = "Circ.1",
  x_lab = "Reporter",
  y_lab = "Circularity (index)",
  file_name = "plot"
)
}
\arguments{
\item{data}{the data that will be input}

\item{x}{string denoting the x parameter (categorical)}

\item{y}{string denoting the y parameter}

\item{x_lab}{string denoting the x label}

\item{y_lab}{string denoting the y label}

\item{file_name}{string denoting how the output should be identified}
}
\value{
a plot
}
\description{
A function to plot
}
\author{
Jones
}
