% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_tables.R
\name{grep_collected}
\alias{grep_collected}
\title{Grep names of dataframes}
\usage{
grep_collected(data, pattern)
}
\arguments{
\item{data}{dataframes}

\item{pattern}{a regular expression that matches the names of
dataframes in a list of data frames}
}
\value{
a list of dataframes that match the specified criteria
}
\description{
Grep Collected
}
\details{
A helper function to pattern-match against the names of the
dataframes.
}
\author{
Jones
}
