% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_tables.R
\name{label_lists}
\alias{label_lists}
\title{Label Lists}
\usage{
label_lists(b, labels)
}
\arguments{
\item{b}{the dataframe list to be labeled}

\item{labels}{the labels}
}
\value{
the dataframe labeled
}
\description{
Label Lists
}
