% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_tables.R
\name{merge_folder}
\alias{merge_folder}
\title{Merge Folder}
\usage{
merge_folder(path, name = "combined")
}
\arguments{
\item{path}{}
}
\value{
one dataframe
}
\description{
Merge all of the tables in a particular folder into one dataframe, write it, and return it.
}
\author{
Jones
}
