% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagej_tables.R
\name{write_list_of_names}
\alias{write_list_of_names}
\title{Create a CSV containing a list of the file names}
\usage{
write_list_of_names(folder = ".", id = "-names")
}
\arguments{
\item{folder}{the path to a directory}

\item{id}{the string appended to describe the document}
}
\value{
a string describing the result
}
\description{
Write List of Names
}
\details{
Write a list of names of the files in a particular folder.
}
\author{
Jones
}
